+++ conans/test/unittests/client/util/files_test.py
@@ -6,7 +6,7 @@ import unittest
 from conans.client.cmd.uploader import compress_files
 from conans.paths import PACKAGE_TGZ_NAME
 from conans.test.utils.test_files import temp_folder
-from conans.util.files import md5sum, mkdir, path_exists, save
+from conans.util.files import md5sum, save
 
 
 class FilesTest(unittest.TestCase):
@@ -44,14 +44,3 @@ class FilesTest(unittest.TestCase):
         md5_b = md5sum(file_path)
 
         self.assertEqual(md5_a, md5_b)
-
-    def test_path_exists(self):
-        """
-        Unit test of path_exists
-        """
-        tmp_dir = temp_folder()
-        tmp_dir = os.path.join(tmp_dir, "WhatEver")
-        new_path = os.path.join(tmp_dir, "CapsDir")
-        mkdir(new_path)
-        self.assertTrue(path_exists(new_path, tmp_dir))
-        self.assertFalse(path_exists(os.path.join(tmp_dir, "capsdir"), tmp_dir))
