+++ conans/tools.py
@@ -20,9 +20,8 @@ from conans.client.tools.scm import *  # pylint: disable=unused-import
 from conans.client.tools.settings import *  # pylint: disable=unused-import
 from conans.client.tools.apple import *
 # Tools form conans.util
-from conans.util.files import _generic_algorithm_sum, load, md5, md5sum, mkdir, relative_dirs, \
-    rmdir, save as files_save, save_append, sha1sum, sha256sum, to_file_bytes, touch
-
+from conans.util.files import _generic_algorithm_sum, load, md5, md5sum, mkdir, rmdir, save as files_save, save_append, sha1sum, sha256sum, to_file_bytes, touch
+from conans.server.utils.files import relative_dirs
 
 # This global variables are intended to store the configuration of the running Conan application
 _global_requester = None
