+++ conans/util/env.py
@@ -10,9 +10,6 @@ import os
 from contextlib import contextmanager
 
 
-# TODO: Rename this file, it is not longer "env_reader" only
-
-
 @contextmanager
 def environment_update(env_vars):
     old_env = dict(os.environ)
